+++ src/test/java/blackjack/domain/ScoreTest.java
@@ -1,4 +1,4 @@
-package blackjacktest.domaintest;
+package blackjack.domain;
 
 import blackjack.domain.Score;
 import blackjack.domain.card.Card;
