+++ src/test/java/blackjack/domain/card/CardDeckTest.java
@@ -1,9 +1,5 @@
-package blackjacktest.domaintest.cardtest;
+package blackjack.domain.card;
 
-import blackjack.domain.card.Card;
-import blackjack.domain.card.CardDeck;
-import blackjack.domain.card.Denomination;
-import blackjack.domain.card.Shape;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
