+++ src/test/java/blackjack/domain/gamer/DealerTest.java
@@ -1,10 +1,9 @@
-package blackjacktest.domaintest.gamertest;
+package blackjack.domain.gamer;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
 import blackjack.domain.card.Denomination;
 import blackjack.domain.card.Shape;
-import blackjack.domain.gamer.Dealer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
