+++ src/test/java/blackjack/domain/gamer/PlayersTest.java
@@ -1,11 +1,7 @@
-package blackjacktest.domaintest.gamertest;
+package blackjack.domain.gamer;
 
-import blackjack.domain.gamer.Player;
-import blackjack.domain.gamer.Players;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.Arrays;
 import java.util.List;
