+++ src/test/java/fuelinjection/RentCompanyTest.java
@@ -1,9 +1,5 @@
-package fuelinjectiontest;
+package fuelinjection;
 
-import fuelinjection.Avante;
-import fuelinjection.K5;
-import fuelinjection.RentCompany;
-import fuelinjection.Sonata;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
